voiceassistance/pyproject.toml
[tool.poetry]
name = "voiceassistance"
version = "0.1.0"
description = "Functionalities developed in the On WHeels project for voice assistance"
authors = ["sophieDataroots <sophie.decoppel@dataroots.io>"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.9.0,<3.12.0"
unidecode = "^1.3.4"
uvicorn = { extras = ["standard"], version = "^0.18.2" }
python-multipart = "^0.0.5"
ffmpeg-python = "^0.2.0"
fastapi = "^0.95.1"
openai = "0.27.6"
flatdict = "4.0.1"
gradio = "3.35.0"
guardrails-ai = {git = "https://github.com/sophieDataroots/guardrails.git"}
py-espeak-ng = "0.1.8"
openai-whisper = {git = "https://github.com/openai/whisper.git"}
tts = {git = "https://github.com/coqui-ai/TTS.git", rev = "37b558ccb99f56d24d0983b1b9480a15334897f7"}
unidic-lite = "^1.0.8"
mecab-python3 = "^1.0.6"
setuptools-rust = "^1.6.0"
pytest = "^7.4.0"
pre-commit = "^3.3.3"
torch = ">=2.0.0, !=2.0.1"

[tool.poetry.group.dev.dependencies]
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
